renamed some files to show that this can be used on both articles and publication pages
