+++ source/js/foundation/template-js-handler/index.js
@@ -11,7 +11,7 @@ import publicationSummaryBar from "./publication-summary-bar.js";
 import continuousScrollingBox from "./continuous-scrolling-box.js";
 import pulseProfileCardHandler from "./pulse-profile-card.js";
 import pulseProfileListFilterHandler from "./pulse-profile-list-filter.js";
-import articleHeroVideoHandler from "./article-hero-video-handler.js";
+import publicationArticleHeroVideoHandler from "./publication-article-hero-video-handler.js";
 import blogHeroVideoHandler from "./blog-hero-video-handler.js";
 /**
  * Bind global event handlers
@@ -27,12 +27,12 @@ export const bindWindowEventHandlers = () => {
  */
 export const bindEventHandlers = () => {
   articleFootLinkHandler();
-  articleHeroVideoHandler();
   audioBlockHandler();
   homepageHandler();
   loadMoreEntriesHandler();
   loopingVideoHandler();
   participatePageDonateHandler();
+  publicationArticleHeroVideoHandler();
   publicationSummaryBar();
   pulseProfileCardHandler();
   pulseProfileListFilterHandler();
